Small reorganizing
