+++ DebugHelper/DebugHelper/Systems/DebugOverlay.cs
@@ -20,7 +20,6 @@ namespace DebugHelper.Systems
         {
             var gameObject = Instantiate(Main.assetBundle.LoadAsset<GameObject>("DebugOverlayCanvas"));
             var component = gameObject.AddComponent<DebugOverlay>();
-            gameObject.AddComponent<DebugCollidersManager>();
             component.rawImage = gameObject.GetComponentInChildren<RawImage>();
         }
